added the owner creation in the project creation

+++ IssueTracker/IssueTrackerAPI/Controllers/ProjectController.cs
@@ -1,8 +1,11 @@
 ﻿using AutoMapper;
 using DataAccess.Data.IData;
 using DataAccess.Models;
+using DataAccess.Utils;
 using FluentValidation.Results;
 using IssueTrackerAPI.Utils;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Models.Request;
 using Validation;
@@ -14,14 +17,17 @@ namespace IssueTrackerAPI.Controllers
     public class ProjectController : ControllerBase
     {
         private readonly IProjectData _projectdb;
+        private readonly IParticipantData _participantdb;
         private readonly Mapper _mapper;
-        public ProjectController(IProjectData projectdb)
+        public ProjectController(IProjectData projectdb, IParticipantData participantdb)
         {
             _projectdb = projectdb;
+            _participantdb = participantdb;
             _mapper = AutoMapperConfig.Config();
         }
 
         [HttpPost("add-project")]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         public async Task<IActionResult> AddProject(ProjectRequest entity)
         {
             var validator = new ProjectValidation();
@@ -32,8 +38,12 @@ namespace IssueTrackerAPI.Controllers
                 return BadRequest(failures);
             }
             var project = _mapper.Map<Project>(entity);
-            await _projectdb.AddAsync(project);
-            return Ok();
+            var res = await _projectdb.AddAsync(project);
+            var userid = User.Claims.FirstOrDefault(x => x.Type.Equals("UserId"));
+            var results = await ParticipantController.CreateOwner(res, Guid.Parse(userid!.Value), _participantdb);
+            if(results)
+                return Ok();
+            return BadRequest();
         }
 
         [HttpGet("getAll-project")]
@@ -48,6 +58,7 @@ namespace IssueTrackerAPI.Controllers
             
         }
         [HttpPut("update-project")]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         public async Task<IActionResult> Update(ProjectRequest entity)
         {
             var validator = new ProjectValidation();
@@ -61,17 +72,37 @@ namespace IssueTrackerAPI.Controllers
             {
                 return BadRequest("Validation error!");
             }
-            var project = _mapper.Map<Project>(entity);
-            await _projectdb.UpdateAsync(project);
-            return Ok();
+            var idclaim = User.Claims.FirstOrDefault(x => x.Type.Equals("UserId"));
+            if (idclaim != null)
+            {
+                var results = await CheckRole.IsOwner(_participantdb,
+                        Guid.Parse(idclaim.Value), entity.Id);
+                if (results == true)
+                {
+                    var project = _mapper.Map<Project>(entity);
+                    await _projectdb.UpdateAsync(project);
+                    return Ok();
+                }
+            }
+            return Unauthorized();
         }
 
         [HttpDelete("delete-project")]
         public async Task<IActionResult> Update(int id)
         {
             if (id <= 0) return BadRequest("Invalid Id!");
-            await _projectdb.DeleteAsync(id);
-            return Ok();
+            var idclaim = User.Claims.FirstOrDefault(x => x.Type.Equals("UserId"));
+            if (idclaim != null)
+            {
+                var results = await CheckRole.IsOwner(_participantdb,
+                        Guid.Parse(idclaim.Value), id);
+                if (results == true)
+                {
+                    await _projectdb.DeleteAsync(id);
+                    return Ok();
+                }
+            }
+            return Unauthorized();
         }
     }
 }
